base refactor go brrrrrr
