+++ azext_edge/tests/edge/checks/test_opcua_checks_unit.py
@@ -34,7 +34,7 @@ def test_check_opcua_by_resource_types(ops_service, mocker, mock_resource_types,
         OpcuaResourceKinds.ASSET_TYPE.value: "azext_edge.edge.providers.check.opcua.evaluate_asset_types",
     }
 
-    assert_check_by_resource_types(ops_service, mocker, mock_resource_types, resource_kinds, eval_lookup)
+    assert_check_by_resource_types(ops_service, mocker, resource_kinds, eval_lookup)
 
 
 @pytest.mark.parametrize("detail_level", ResourceOutputDetailLevel.list())
